leo_chat/requirements.txt
# Web UI
streamlit>=1.28.0

# Data processing
beautifulsoup4>=4.12.0
aiohttp>=3.8.5
tqdm>=4.65.0

# Machine Learning
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# Add these to requirements.txt
python-docx>=0.8.11  # For .docx files
PyPDF2>=3.0.0  # For .pdf files 